+++ gsa/src/web/pages/policies/component.js
@@ -32,7 +32,6 @@ import {selectSaveId} from 'gmp/utils/id';
 import {YES_VALUE, NO_VALUE} from 'gmp/parser';
 
 import {
-  ospScannersFilter,
   OPENVAS_DEFAULT_SCANNER_ID,
   OPENVAS_SCANNER_TYPE,
   GREENBONE_SENSOR_SCANNER_TYPE,
@@ -103,29 +102,23 @@ class PolicyComponent extends React.Component {
     this.handleSavePolicyFamily = this.handleSavePolicyFamily.bind(this);
     this.handleSavePolicyNvt = this.handleSavePolicyNvt.bind(this);
     this.openCreatePolicyDialog = this.openCreatePolicyDialog.bind(this);
-    this.handleCloseCreatePolicyDialog = this.handleCloseCreatePolicyDialog.bind(
-      this,
-    );
+    this.handleCloseCreatePolicyDialog =
+      this.handleCloseCreatePolicyDialog.bind(this);
     this.openEditPolicyDialog = this.openEditPolicyDialog.bind(this);
-    this.handleCloseEditPolicyDialog = this.handleCloseEditPolicyDialog.bind(
-      this,
-    );
-    this.openEditPolicyFamilyDialog = this.openEditPolicyFamilyDialog.bind(
-      this,
-    );
-    this.handleCloseEditPolicyFamilyDialog = this.handleCloseEditPolicyFamilyDialog.bind(
-      this,
-    );
+    this.handleCloseEditPolicyDialog =
+      this.handleCloseEditPolicyDialog.bind(this);
+    this.openEditPolicyFamilyDialog =
+      this.openEditPolicyFamilyDialog.bind(this);
+    this.handleCloseEditPolicyFamilyDialog =
+      this.handleCloseEditPolicyFamilyDialog.bind(this);
     this.openEditNvtDetailsDialog = this.openEditNvtDetailsDialog.bind(this);
-    this.handleCloseEditNvtDetailsDialog = this.handleCloseEditNvtDetailsDialog.bind(
-      this,
-    );
+    this.handleCloseEditNvtDetailsDialog =
+      this.handleCloseEditNvtDetailsDialog.bind(this);
     this.openImportDialog = this.openImportDialog.bind(this);
     this.handleCloseImportDialog = this.handleCloseImportDialog.bind(this);
     this.openCreateAuditDialog = this.openCreateAuditDialog.bind(this);
-    this.handleCloseCreateAuditDialog = this.handleCloseCreateAuditDialog.bind(
-      this,
-    );
+    this.handleCloseCreateAuditDialog =
+      this.handleCloseCreateAuditDialog.bind(this);
     this.handleSaveAudit = this.handleSaveAudit.bind(this);
     this.handleAlertCreated = this.handleAlertCreated.bind(this);
     this.handleScheduleCreated = this.handleScheduleCreated.bind(this);
@@ -532,8 +525,7 @@ class PolicyComponent extends React.Component {
     return gmp.scanners
       .getAll()
       .then(response => {
-        let {data: scanners} = response;
-        scanners = scanners.filter(ospScannersFilter);
+        const {data: scanners} = response;
         return {
           scanners,
           scannerId: selectSaveId(scanners),