+++ gsa/src/web/pages/scanconfigs/__tests__/details.js
@@ -19,12 +19,9 @@ import React from 'react';
 
 import Capabilities from 'gmp/capabilities/capabilities';
 
-import ScanConfig, {
-  OPENVAS_SCAN_CONFIG_TYPE,
-  OSP_SCAN_CONFIG_TYPE,
-} from 'gmp/models/scanconfig';
+import ScanConfig, {OPENVAS_SCAN_CONFIG_TYPE} from 'gmp/models/scanconfig';
 
-import {OPENVAS_SCANNER_TYPE, OSP_SCANNER_TYPE} from 'gmp/models/scanner';
+import {OPENVAS_SCANNER_TYPE} from 'gmp/models/scanner';
 
 import {rendererWith} from 'web/utils/testing';
 
@@ -38,7 +35,10 @@ describe('Scan Config Details tests', () => {
       scanner: {name: 'scanner1', id: '42', type: OPENVAS_SCANNER_TYPE},
       type: OPENVAS_SCAN_CONFIG_TYPE,
       tasks: {
-        task: [{id: '1234', name: 'task1'}, {id: '5678', name: 'task2'}],
+        task: [
+          {id: '1234', name: 'task1'},
+          {id: '5678', name: 'task2'},
+        ],
       },
     });
     const caps = new Capabilities(['everything']);
@@ -59,36 +59,5 @@ describe('Scan Config Details tests', () => {
     expect(detailslinks[1]).toHaveAttribute('href', '/task/5678');
 
     expect(element).not.toHaveTextContent('scanner');
-  });
-
-  test('should render details for osp config', () => {
-    const config = ScanConfig.fromElement({
-      name: 'foo',
-      comment: 'bar',
-      scanner: {_id: '42', name: 'scanner1', type: OSP_SCANNER_TYPE},
-      type: OSP_SCAN_CONFIG_TYPE,
-      tasks: {
-        task: [{id: '1234', name: 'task1'}, {id: '5678', name: 'task2'}],
-      },
-    });
-    const caps = new Capabilities(['everything']);
-
-    const {render} = rendererWith({capabilities: caps, router: true});
-
-    const {element, getAllByTestId} = render(<Details entity={config} />);
-
-    expect(element).toHaveTextContent('bar');
-
-    const detailslinks = getAllByTestId('details-link');
-
-    expect(element).toHaveTextContent('task1');
-    expect(detailslinks[1]).toHaveAttribute('href', '/task/1234');
-
-    expect(element).toHaveTextContent('task2');
-    expect(detailslinks[2]).toHaveAttribute('href', '/task/5678');
-
-    expect(element).toHaveTextContent('Scanner');
-    expect(detailslinks[0]).toHaveAttribute('href', '/scanner/42');
-    // TODO test if scanner name is displayed correctly
   });
 });